backend/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart

